Move inherited seeding files
