+++ database/seed/inheritedSeed/seed-db.js
@@ -1,4 +1,4 @@
-const db = require('../connect.js');
+const db = require('../../connect.js');
 const seed = require('./seed-functions.js');
 const data = require('./seed-data.json');
